Comprehensive list of documented constants and functions

Constants

BiblicalHebrew.begadkepat — Constant

Vector of the begadkepat letters as Chars.

BiblicalHebrew.bet_ch — Constant

Unicode codepoint for bet as a Char.

BiblicalHebrew.dagesh_ch — Constant

Unicode codepoint for dagesh as a Char.

BiblicalHebrew.dalet_ch — Constant

Unicode codepoint for dalet as a Char.

BiblicalHebrew.gimel_ch — Constant

Unicode codepoint for gimel as a Char.

BiblicalHebrew.kaph_ch — Constant

Unicode codepoint for kaph as a Char.

BiblicalHebrew.kaph_final_ch — Constant

Unicode codepoint for final kaph as a Char.

BiblicalHebrew.mappiq_ch — Constant

Unicode codepoint for mappiq as a Char.

BiblicalHebrew.pe_ch — Constant

Unicode codepoint for pe as a Char.

BiblicalHebrew.pe_final_ch — Constant

Unicode codepoint for final pe as a Char.

BiblicalHebrew.shin_dot_ch — Constant

Unicode codepoint for shin dot as a Char.

BiblicalHebrew.sin_dot_ch — Constant

Unicode codepoint for sin dot as a Char.

BiblicalHebrew.tav_ch — Constant

Unicode codepoint for tav as a Char.

Functions

BiblicalHebrew.codept — Method

Find Unicode codepoint for a hexadecimal string.

BiblicalHebrew.codept_name — Method

Look up Unicode summary name for a codepoint in the Hebrew range of Unicode.

BiblicalHebrew.codept_split — Method

Split string s into substrings separated by character c; if keep is true, also maintain c as a string in the resulting list.

BiblicalHebrew.dagesh — Method

Add dagesh to string s.

dagesh(s)

BiblicalHebrew.gershe — Method

Append gershe to string s.

BiblicalHebrew.hebrew_chars — Method

Compose a list of all defined characters in the Hebrew range of Unicode as Chars, sorted in Unicode order.

BiblicalHebrew.hex — Method

Compose hexadecimal string for integer value of ch.

BiblicalHebrew.hex — Method

Find hexadecimal string for an integer.

BiblicalHebrew.hiriq — Method

Add hiriq to string s.

BiblicalHebrew.holam — Method

Add holam to string s.

BiblicalHebrew.int — Method

Find integer value of a hexadecimal string.

BiblicalHebrew.is_alphabetic — Method

True if a Char is a vowel point, a consonant, or one of the three Unicode combining characters that are part of writing consonantal values.

is_alphabetic(c)

BiblicalHebrew.is_begadkepat — Method

True if initial consonant of s is a consonant in the begadkepat.

is_begadkepat(s)

BiblicalHebrew.is_consonant — Method

True if Char is a Hebrew consonant.

is_consonant(c)

BiblicalHebrew.is_guttural — Method

True if String begins with a guttural.

is_guttural(s)

BiblicalHebrew.is_vowelpoint — Method

True if Char is a Hebrew vowel point.

is_vowelpoint(c)

BiblicalHebrew.mappiq — Method

Add mappiq to string s.

BiblicalHebrew.maqaf_join — Method

Connect a list of tokens together with maqaf.

BiblicalHebrew.metheg — Method

Add metheg to string s.

BiblicalHebrew.ole — Method

Add accent ole to string s.

BiblicalHebrew.patah — Method

Add patah to string s.

BiblicalHebrew.qamats — Method

Add qamats to string s.

BiblicalHebrew.qubbuts — Method

Add qubbuts to string s.

BiblicalHebrew.rm_accents — Method

Remove all accents and cantillation marks from a string, leaving only consonants, vowel points and dagesh/mappiq (the same code point in Unicode).

rm_accents(s)

BiblicalHebrew.seghol — Method

Add seghol to string s.

BiblicalHebrew.sheva — Method

Add sheva to string s.

BiblicalHebrew.tokenize_string — Function

Tokenize a string s where s has already isolated the token separators maqaf and soph pasuq. This can be accomplished by feeding s to the tokenize function with a HebrewOrthography.

BiblicalHebrew.tsere — Method

Add tsere to string s.

BiblicalHebrew.unpointed — Method

Reduce a string to consonantal representation by removing all pointing and accents.

unpointed(s)

BiblicalHebrew.validcps — Method

Compose a list of all defined codepoints in the Hebrew range of Unicode sorted in codepoint order.

BiblicalHebrew.ws_chars — Method

Whitespace characters accepted in this orthography

Orthography.codepoints — Method

Implement Orthography's codepoints functions for HebrewOrthography.

codepoints(ortho)

Orthography.tokenize — Method

Implement Orthography's tokenize function for HebrewOrthography.

tokenize(s, o)

Orthography.tokentypes — Method

Implement Orthography's tokentypes functions for HebrewOrthography.

tokentypes(ortho)

Orthography.validstring — Method

Implement validstring function for HebrewOrthography. True if every code point in s appears in the HebrewOrthography's values for codepoints.