Niqqud

The Unicode specification defines combining characters for the Tiberian vowel points (niqqud).

The BiblicalHebrew package defines character constants that can be used to refer to these values (see the complete list of constants in the section "APIs for all constants and functions" of this documentation. It also offers a parallel suite of functions for composing strings with the vowel points.

For example, add a qamats to an aleph:

pointed = BiblicalHebrew.qamats("א")
Markdown.parse("> pointed = " * pointed)

pointed = אָ

You can of course chain these functions. To compose hateph vowels, for example, use the sheva function, then a function for a vowel:

hateph_seghol = BiblicalHebrew.sheva("א") |> BiblicalHebrew.seghol
Markdown.parse("> hateph_seghol is " * hateph_seghol)

hateph_seghol is אְֶ

There is also a suite of functions for adding accents and other marks, such as metheg.

qamats_metheg = BiblicalHebrew.qamats("א") |> BiblicalHebrew.metheg
Markdown.parse("> qamats_metheg is " * qamats_metheg)

qamats_metheg is אָֽ

Functions for adding vowel points to strings

BiblicalHebrew.qamats — Function

Add qamats to string s.

BiblicalHebrew.patah — Function

Add patah to string s.

BiblicalHebrew.seghol — Function

Add seghol to string s.

BiblicalHebrew.hiriq — Function

Add hiriq to string s.

BiblicalHebrew.tsere — Function

Add tsere to string s.

BiblicalHebrew.qubbuts — Function

Add qubbuts to string s.

BiblicalHebrew.holam — Function

Add holam to string s.

BiblicalHebrew.sheva — Function

Add sheva to string s.

Other accent marks

BiblicalHebrew.metheg — Function

Add metheg to string s.

BiblicalHebrew.ole — Function

Add accent ole to string s.

BiblicalHebrew.mappiq — Function

Add mappiq to string s.

BiblicalHebrew.gershe — Function

Append gershe to string s.

BiblicalHebrew.maqaf_join — Function

Connect a list of tokens together with maqaf.