Generic utilities for working with strings and characters

Even in the twenty-first century, it can be awkward in some environments to display codepoints in the Hebrew range of Unicode. The package includes a few shortcuts for viewing different representations of codepoints that (although generic) are sometimes especially useful when working with fully pointed Hebrew text.

BiblicalHebrew.codept — Function

Find Unicode codepoint for a hexadecimal string.

BiblicalHebrew.hex — Function

Find hexadecimal string for an integer.

Compose hexadecimal string for integer value of ch.

BiblicalHebrew.int — Function

Find integer value of a hexadecimal string.

BiblicalHebrew.codept_split — Function

Split string s into substrings separated by character c; if keep is true, also maintain c as a string in the resulting list.